requirements.txt
faker
apache-beam
kafka-python